+++ libs/ledger-live-common/src/families/evm/react.ts
@@ -1,5 +1,5 @@
 import { getGasTracker } from "@ledgerhq/coin-evm/api/gasTracker/index";
-import type { GasOptions, Transaction } from "@ledgerhq/coin-evm/types";
+import type { GasOptions, Transaction } from "@ledgerhq/coin-evm/types/index";
 import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 import { useEffect, useMemo, useState } from "react";
